+++ internal/ui/pages/admin/middleware/auth.go
@@ -1,4 +1,4 @@
-package adminmiddleware
+package middleware
 
 import (
 	"context"
